+++ frontend/src/helpers/countPrices.js
@@ -1,97 +1,97 @@
-export const countPrices = (numOfFilials, duration) => {
+export const countPrices = (numOfBranchs, duration) => {
     let resultPrice = 0;
     let oldPrice = 0;
-    if (numOfFilials < 10){
+    if (numOfBranchs < 10) {
         return {oldPrice: duration, resultPrice: duration, sale: '0%'};
     }
-    if (numOfFilials >= 10 && numOfFilials < 15) {
+    if (numOfBranchs >= 10 && numOfBranchs < 15) {
         // скидка 5%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 95);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 95);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '5%'};
     }
-    if (numOfFilials >= 15 && numOfFilials < 25) {
+    if (numOfBranchs >= 15 && numOfBranchs < 25) {
         // скидка 10%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 90);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 90);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '10%'};
     }
-    if (numOfFilials >= 25 && numOfFilials < 35) {
+    if (numOfBranchs >= 25 && numOfBranchs < 35) {
         // скидка 15%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 85);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 85);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '15%'};
     }
-    if (numOfFilials >= 35 && numOfFilials < 45) {
+    if (numOfBranchs >= 35 && numOfBranchs < 45) {
         // скидка 20%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 80);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 80);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '20%'};
     }
-    if (numOfFilials >= 45 && numOfFilials < 55) {
+    if (numOfBranchs >= 45 && numOfBranchs < 55) {
         // скидка 25%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 75);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 75);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '25%'};
     }
-    if (numOfFilials >= 55 && numOfFilials < 65) {
+    if (numOfBranchs >= 55 && numOfBranchs < 65) {
         // скидка 30%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 70);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 70);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '30%'};
     }
-    if (numOfFilials >= 65 && numOfFilials < 75) {
+    if (numOfBranchs >= 65 && numOfBranchs < 75) {
         // скидка 35%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 65);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 65);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '35%'};
     }
-    if (numOfFilials >= 75 && numOfFilials < 85) {
+    if (numOfBranchs >= 75 && numOfBranchs < 85) {
         // скидка 40%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 60);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 60);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '40%'};
     }
-    if (numOfFilials >= 85 && numOfFilials < 95) {
+    if (numOfBranchs >= 85 && numOfBranchs < 95) {
         // скидка 45%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 55);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 55);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '45%'};
     }
-    if (numOfFilials >= 95 && numOfFilials < 105) {
+    if (numOfBranchs >= 95 && numOfBranchs < 105) {
         // скидка 50%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 50);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 50);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '50%'};
     }
-    if (numOfFilials >= 105 && numOfFilials < 115) {
+    if (numOfBranchs >= 105 && numOfBranchs < 115) {
         // скидка 55%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 45);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 45);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '55%'};
     }
-    if (numOfFilials >= 115 && numOfFilials < 125) {
+    if (numOfBranchs >= 115 && numOfBranchs < 125) {
         // скидка 60%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 40);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 40);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '60%'};
     }
-    if (numOfFilials >= 125 && numOfFilials < 135) {
+    if (numOfBranchs >= 125 && numOfBranchs < 135) {
         // скидка 65%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 35);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 35);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '65%'};
     }
-    if (numOfFilials >= 135 && numOfFilials < 145) {
+    if (numOfBranchs >= 135 && numOfBranchs < 145) {
         // скидка 70%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 30);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 30);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '70%'};
     }
-    if (numOfFilials >= 145) {
+    if (numOfBranchs >= 145) {
         // скидка 75%
-        oldPrice = duration * numOfFilials
-        resultPrice = Math.floor(((duration * numOfFilials) / 100) * 25);
+        oldPrice = duration * numOfBranchs
+        resultPrice = Math.floor(((duration * numOfBranchs) / 100) * 25);
         return {oldPrice: oldPrice, resultPrice: resultPrice, sale: '75%'};
     }
 }